requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
Django==3.0
django-background-tasks==1.2.5
django-compat==1.0.15
entrypoints==0.3
flake8==3.7.9
idna==2.8
mccabe==0.6.1
multidict==4.7.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
six==1.13.0
slackclient==2.5.0
sqlparse==0.3.0
yarl==1.4.2
